builder-cleanup: keep latest packages in qubes-packages-mirror-repo
Previous version relied on hardlinks to packages in
qubes-packages-mirror-repo to check their usage, but it doesn't work
because those hardlinks are broken when package is moved out of chroot
and later also when it's signed. This resulted in
qubes-packages-mirror-repo being cleaned from all the packages, breaking
future builds (missing build deps).

Fix this by using repomanage as for other directories.

Note the previous method is still fine for qubes-src/*/pkgs, because:
a) those packages are not used as build dependencies
b) hardlinks are not broken here, because this is the place where
packages gets signed